1) Removed mass compensation 2) Add extra padding to cluster subwindows 3) Fix binning mismatch between subwindow and pdf lookup's 4) Saved endtime for each window for future comparison 5) Add more debug statements (including pe time stamps) 6) Allow for fractional seeding from coincidence within a window

+++ include/Pulse_Fitting.h
@@ -197,9 +197,9 @@ class Pulse_Fitting {
         
         std::vector<std::vector<double>> generatePDFLookup(const std::vector<double>& xCenters); // cached shifted PDFs
 
-        std::vector<int> findGradientPeaks(const std::vector<int>& hist, double threshold);
+        std::vector<double> findGradientPeaks(const std::vector<int>& hist, double threshold);
           
-        std::vector<int> findCoincidenceSeeds(double startUs, double endUs, double binWidthUs, int pre_bins) const;
+        std::vector<double> findCoincidenceSeeds(double startUs, double endUs, double binWidthUs, int pre_bins) const;
 
         bool fitPulses(const std::vector<int>& hist, const std::vector<double>& xCenters,
                     const std::vector<std::vector<double>>& pdfLookup,
